bdaproj/dhef_project/requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
pandas==2.1.0
numpy==1.23.5
scikit-learn==1.3.0
imbalanced-learn==0.11.0
matplotlib==3.7.2
seaborn==0.12.2
streamlit==1.28.0
plotly==5.17.0
findspark==2.0.1
# river removed — using built-in pure-Python ADWIN (detector/adwin.py)

